S11P12D210-224 <feat> [be] 라이브강의 id 타입 수정, 날짜 타입 수정

See merge request s11-webmobile4-sub2/S11P12D210!144

+++ backend/src/main/java/com/yoga/backend/mypage/livelectures/dto/LiveLectureResponseDto.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Getter
 public class LiveLectureResponseDto {
 
-    private Integer liveId; // 화상 강의 ID
+    private Long liveId; // 화상 강의 ID
     private Long regDate; // 강의 등록 시간
 
     private int userId; // 강사 ID
